fix bugs and modify

+++ src/main/java/servlet/DeleteFile.java
@@ -9,10 +9,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.File;
 import java.io.IOException;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
+import java.sql.*;
 
 @WebServlet("/pages/deleteFile")
 public class DeleteFile extends HttpServlet {
@@ -46,21 +43,41 @@ public class DeleteFile extends HttpServlet {
     }
 
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String userID = request.getParameter("user_id");
-        String fileName = request.getParameter("file_name");
-
-        // 获取上传的文件对象
-        String uploadPath = request.getServletContext().getRealPath("./") + File.separator + UPLOAD_DIRECTORY;
-        File file = new File(uploadPath + File.separator + fileName);
+        //验证用户是否登录
+        String user_id = request.getSession().getAttribute("user_id").toString();
+        if(user_id == null) {
+            // 发送401 Unauthorized错误
+            response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "User not logged in.");
+            return;
+        }
 
         //删除文件
         try {
-            ps = conn.prepareStatement("DELETE FROM files WHERE user_id = ? AND file_name = ?");
-            ps.setString(1, userID);
-            ps.setString(2, fileName);
+            String file_id = request.getParameter("file_id");
+            //查询文件
+            ps = conn.prepareStatement("SELECT file_name FROM files WHERE file_id = ?");
+            ps.setString(1, file_id);
+            ResultSet rs = ps.executeQuery();
+
+            if(!rs.next()){
+                response.sendError(404, "File not found.");
+                return;
+            }
+
+            String fileName = rs.getString("file_name");
+
+            //删除数据库中的记录
+            ps = conn.prepareStatement("DELETE FROM files WHERE user_id = ? AND file_id = ?");
+            ps.setString(1, user_id);
+            ps.setString(2, file_id);
             ps.executeUpdate();
+
+            // 删除文件
+            String uploadPath = request.getServletContext().getRealPath("./") + File.separator + UPLOAD_DIRECTORY + File.separator + user_id;
+            File file = new File(uploadPath + File.separator + fileName);
             if(file.exists()) file.delete();
-            System.out.println("delete file: " + fileName + ", from user: " + userID);
+
+            System.out.println("delete file: " + fileName + ", from user: " + user_id);
         } catch (SQLException e) {
             e.printStackTrace();
             response.sendError(500, "Failed to delete file from database.");
